refactor(users): share access control options
